% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5mu.R
\name{writeH5MU}
\alias{writeH5MU}
\title{Save a \code{\linkS4class{MultiAssayExperiment}} to an .h5mu file.}
\usage{
writeH5MU(object, file, overwrite)
}
\arguments{
\item{object}{A \code{\linkS4class{MultiAssayExperiment}}.}

\item{file}{Name of the file to save to.}

\item{overwrite}{Currently unused.}
}
\value{
NULL, invisibly
}
\description{
Note that NA values are not supported by HDF5, and therefore by h5mu. The behavior of this
function if NAs are present is undefined.
}
\examples{
data(miniACC, package="MultiAssayExperiment")
writeH5MU(miniACC, "miniacc.h5mu")

}
