% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5mu.R
\name{WriteH5AD}
\alias{WriteH5AD}
\alias{WriteH5AD,Matrix_OR_DelayedMatrix,H5IdComponent-method}
\alias{WriteH5AD,SummarizedExperiment,H5IdComponent-method}
\alias{WriteH5AD,SingleCellExperiment,H5IdComponent-method}
\alias{WriteH5AD,ANY,H5IdComponent-method}
\alias{WriteH5AD,ANY,character-method}
\title{Save an experiment to an .h5ad file.}
\usage{
WriteH5AD(object, file, overwrite = TRUE, ...)

\S4method{WriteH5AD}{Matrix_OR_DelayedMatrix,H5IdComponent}(object, file, overwrite, write_dimnames = TRUE)

\S4method{WriteH5AD}{SummarizedExperiment,H5IdComponent}(object, file, overwrite)

\S4method{WriteH5AD}{SingleCellExperiment,H5IdComponent}(object, file, overwrite)

\S4method{WriteH5AD}{ANY,H5IdComponent}(object, file, overwrite)

\S4method{WriteH5AD}{ANY,character}(object, file, overwrite)
}
\arguments{
\item{object}{The object to save.}

\item{file}{Name of the file to save to.}

\item{overwrite}{Currently unused.}

\item{...}{additional arguments.}

\item{write_dimnames}{Whether to export colnames and rownames.}
}
\value{
NULL, invisibly
}
\description{
Note that NA values are not supported by HDF5, and therefore by h5ad. The behavior of this
function if NAs are present is undefined.
}
\examples{
WriteH5AD(MultiAssayExperiment::miniACC[[1]], "miniacc.h5ad")

}
