% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5mu.R
\name{writeH5AD}
\alias{writeH5AD}
\title{Save an experiment to an .h5ad file.}
\usage{
writeH5AD(object, file, overwrite)
}
\arguments{
\item{object}{The object to save.}

\item{file}{Name of the file to save to.}

\item{overwrite}{Currently unused.}
}
\value{
NULL, invisibly
}
\description{
Note that NA values are not supported by HDF5, and therefore by h5ad. The behavior of this
function if NAs are present is undefined.
}
\examples{
data(miniACC, package="MultiAssayExperiment")
writeH5AD(miniACC[[1]], "miniacc.h5ad")

}
