% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5mu.R
\name{WriteH5AD,ANY,character-method}
\alias{WriteH5AD,ANY,character-method}
\title{Save an experiment to an .h5ad file.}
\usage{
\S4method{WriteH5AD}{ANY,character}(object, file, overwrite)
}
\arguments{
\item{object}{The object to save.}

\item{file}{Name of the file to save to.}

\item{overwrite}{Currently unused.}
}
\description{
Note that NA values are not supported by HDF5, and therefore by h5ad. The behavior of this
function if NAs are present is undefined.
}
