% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_h5mu.R
\name{WriteH5MU}
\alias{WriteH5MU}
\alias{WriteH5MU,MultiAssayExperiment,character-method}
\title{Save a \code{\linkS4class{MultiAssayExperiment}} to an .h5mu file.}
\usage{
WriteH5MU(object, file, overwrite = TRUE)

\S4method{WriteH5MU}{MultiAssayExperiment,character}(object, file, overwrite = TRUE)
}
\arguments{
\item{object}{A \code{\linkS4class{MultiAssayExperiment}}.}

\item{file}{Name of the file to save to.}

\item{overwrite}{Currently unused.}
}
\value{
NULL, invisibly
}
\description{
Note that NA values are not supported by HDF5, and therefore by h5mu. The behavior of this
function if NAs are present is undefined.
}
\examples{
data(miniACC, package="MultiAssayExperiment")
WriteH5MU(miniACC, "miniacc.h5mu")

}
